Development 201707 (#112)
* update for ui

* update for new ui

* update for new ui

* add todo for table api limitation

* updated

* updated

* updated

* updated

* update for new ui

* updated

* updated

* fix mongo string and add lost status when active connector not found.

* updated

* add logging api end and connectType to connector.class

* test and fix connector.class default value features

* update for new ui

* update for new ui

* integrated with flink rest API for job update, delete, and status.

* when there is no schema subject, use topic for instead

+++ src/main/java/com/datafibers/util/MongoAdminClient.java
@@ -1,25 +1,48 @@
 package com.datafibers.util;
 
 import com.mongodb.MongoClient;
+import com.mongodb.client.MongoCollection;
 import com.mongodb.client.MongoCursor;
 import com.mongodb.client.MongoDatabase;
 import com.mongodb.client.MongoIterable;
+import static com.mongodb.client.model.Filters.eq;
+
+import io.vertx.core.json.JsonObject;
+import org.bson.Document;
 
 /**
- * Created by DUW3 on 11/17/2016.
+ * MongoDB Client
  */
 public class MongoAdminClient {
     private MongoClient mongoClient;
     private MongoDatabase database;
+    private MongoCollection<Document> collection;
 
 
     public MongoAdminClient(String hostname, int port, String database) {
         this.mongoClient = new MongoClient(hostname, port );
         this.database = this.mongoClient.getDatabase(database);
+    }
 
+    public MongoAdminClient(String hostname, String port, String database) {
+        this.mongoClient = new MongoClient(hostname, Integer.parseInt(port));
+        this.database = this.mongoClient.getDatabase(database);
+    }
+
+    public MongoAdminClient(String hostname, String port, String database, String collection) {
+        this.mongoClient = new MongoClient(hostname, Integer.parseInt(port));
+        this.database = this.mongoClient.getDatabase(database);
+        this.collection = this.database.getCollection(collection);
+    }
+
+    public MongoAdminClient truncateCollection(String colName) {
+        if(collectionExists(colName))
+        this.database.getCollection(colName).deleteMany(new Document());
+        return this;
     }
 
     public MongoAdminClient dropCollection(String colName) {
+        if(collectionExists(colName))
         this.database.getCollection(colName).drop();
         return this;
     }
@@ -27,6 +50,12 @@ public class MongoAdminClient {
     public MongoAdminClient createCollection(String colName) {
         if(!collectionExists(colName))
         this.database.createCollection(colName);
+        return this;
+    }
+
+    public MongoAdminClient useCollection(String colName) {
+        if(!collectionExists(colName))
+            this.collection = this.database.getCollection(colName);
         return this;
     }
 
@@ -45,5 +74,10 @@ public class MongoAdminClient {
         }
 
         return false;
+    }
+
+    public String findConnectorClassName(String connectorType) {
+        String connectorClass = this.collection.find(eq("connectorType", connectorType)).first().toJson();
+        return new JsonObject(connectorClass).getString("class");
     }
 }
